fix warnings on macos

+++ examples/C/IpIntelligence/MemIpi.c
@@ -94,8 +94,8 @@ typedef struct t_thread_memory_state {
  */
 void printLoadBar(memoryThreadState* state) {
 	int i;
-	int full = state->count / state->main->progress;
-	int empty = (state->main->ipAddressesCount - state->count) /
+	long full = state->count / state->main->progress;
+	long empty = (state->main->ipAddressesCount - state->count) /
 		state->main->progress;
 	printf("\r\t[");
 	for (i = 0; i < full; i++) {
